Fixed test_execute_trade_uniswap
